Move position services into graphical package
