+++ src/app/shared/keml/graphical/msg/msg.component.ts
@@ -6,7 +6,7 @@ import {
   OnInit, Output, AfterViewInit,
 } from '@angular/core';
 import {InformationLink, Message, SendMessage, ReceiveMessage, Information} from "@app/shared/keml/core/msg-info";
-import {LayoutingService} from "@app/features/editor/services/layouting.service";
+import {LayoutingService} from "@app/shared/keml/graphical/layouting.service";
 import {ArrowType} from "@app/shared/keml/graphical/helper/arrow-styling/arrow.models";
 import { NewInfoComponent } from '@app/shared/keml/graphical/new-info/new-info.component';
 import { NgIf, NgFor } from '@angular/common';
